feat: Complete Phase 2 - Move factories and registries
