+++ src/fapilog/__init__.py
@@ -13,10 +13,10 @@ import structlog
 # They occur when test frameworks don't properly await FastAPI shutdown coroutines.
 # This is normal behavior and doesn't affect functionality.
 from ._internal.context import get_trace_id as get_current_trace_id
-from ._internal.processor_registry import register_processor
-from ._internal.sink_registry import SinkRegistry, register_sink
 from .bootstrap import configure_logging, create_logger
 from .container import LoggingContainer
+from .core.registries.processor_registry import register_processor
+from .core.registries.sink_registry import SinkRegistry, register_sink
 from .settings import LoggingSettings
 from .sinks import Sink
