+++ src/fapilog/core/managers/sink_manager.py
@@ -18,13 +18,13 @@ from typing import Any, List, Optional
 
 from ..._internal.error_handling import handle_configuration_error
 from ..._internal.queue_worker import QueueWorker
-from ..._internal.sink_factory import create_custom_sink_from_uri
 from ...exceptions import SinkConfigurationError, SinkErrorContextBuilder
 from ...settings import LoggingSettings
 from ...sinks.base import Sink
 from ...sinks.file import create_file_sink_from_uri
 from ...sinks.loki import create_loki_sink_from_uri
 from ...sinks.stdout import StdoutSink
+from ..factories.sink_factory import create_custom_sink_from_uri
 
 
 class SinkManager:
