+++ src/fapilog/core/registries/component_registry.py
