+++ tests/test_sink_registry.py
@@ -4,14 +4,14 @@ from typing import Any, Dict
 
 import pytest
 
-from fapilog._internal.sink_factory import (
+from fapilog._internal.sink_registry import SinkRegistry, register_sink
+from fapilog.bootstrap import configure_logging
+from fapilog.core.factories.sink_factory import (
     SinkConfigurationError,
     _convert_parameter_value,
     _parse_uri_parameters,
     create_custom_sink_from_uri,
 )
-from fapilog._internal.sink_registry import SinkRegistry, register_sink
-from fapilog.bootstrap import configure_logging
 from fapilog.settings import LoggingSettings
 from fapilog.sinks import Sink
